Fix Google auth for organizations
For Google OAuth, flow could not be started from a subdomain URL since each URL needs to be explicitly listed as Authorized domains inside the Google Project (wildcards not allowed and edits cannot be done via API). So instead, a quick redirect happens to a URL without a subdomain and the organization ID is passed as query param. Then Google OAuth client is started programatically when the component loads. Popup could not be used because browser would block it if it was done programatically. Redirect method was used so I had to set up Cloud functions endpoints for the redirects, then it responds with a redirect to the correct subdomain with a custom token and then finally the authentication is done inside this component with this custom token.

Also, the regular react-google-oauth package could not be used because it did not allow sending a state in the Oauth flow. We needed a state to be passed to the back end in order to determine which organization we were doing the OAuth flow for, then redirect back to the correct subdomain.